add sweet alert and last update survey

+++ src/pages/AdminComponent/UserList.jsx
@@ -6,20 +6,48 @@ import { jwtDecode } from "jwt-decode";
 import edit from "../../assets/edit.svg";
 import deleted from "../../assets/delete.svg";
 import ModalAdd from "./ModalAdd";
+import Swal from "sweetalert2";
 
 const UserList = () => {
   const accessToken = localStorage.getItem("accessToken");
   const [token, setToken] = useState(false);
   const dispatch = useDispatch();
   const { data } = useSelector((state) => state.adminReducers);
+  const [users, setUsers] = useState([]);
+
   useEffect(() => {
     dispatch(getAllUser());
     setToken(jwtDecode(accessToken));
   }, [dispatch, accessToken]);
 
+  useEffect(() => {
+    if (data && data.user) {
+      setUsers(data.user);
+    }
+  }, [data]);
+
   const handleDeleteUser = (userId) => {
-    dispatch(deleteUser(userId));
-    window.location.reload();
+    Swal.fire({
+      title: "Are you sure?",
+      text: "You won't be able to revert this!",
+      icon: "warning",
+      showCancelButton: true,
+      confirmButtonColor: "#3085d6",
+      cancelButtonColor: "#d33",
+      confirmButtonText: "Yes, delete it!",
+    }).then((result) => {
+      if (result.isConfirmed) {
+        dispatch(deleteUser(userId));
+        setUsers(users.filter((user) => user.id !== userId)); // Update local state
+        Swal.fire({
+          title: "Deleted!",
+          text: "Your file has been deleted.",
+          icon: "success",
+          showConfirmButton: false,
+          timer: 1500,
+        });
+      }
+    });
   };
 
   return (
@@ -40,30 +68,29 @@ const UserList = () => {
             </tr>
           </thead>
           <tbody>
-            {data &&
-              data.user?.map((user, no) => (
-                <tr key={user.id}>
-                  <td>{no + 1}</td>
-                  <td>
-                    <p className="font-semibold">{user.username}</p>
-                  </td>
-                  <td>{user.role}</td>
-                  <td className="px-6 py-4">
-                    <div className="flex gap-2">
-                      <button className="font-medium text-blue-600 dark:text-blue-500 hover:underline">
-                        <img src={edit} alt="" />
-                      </button>
+            {users.map((user, index) => (
+              <tr key={user.id}>
+                <td>{index + 1}</td>
+                <td>
+                  <p className="font-semibold">{user.username}</p>
+                </td>
+                <td>{user.role}</td>
+                <td className="px-6 py-4">
+                  <div className="flex gap-2">
+                    {/* <button className="font-medium text-blue-600 dark:text-blue-500 hover:underline">
+                      <img src={edit} alt="" />
+                    </button> */}
 
-                      <button
-                        onClick={() => handleDeleteUser(user.id)}
-                        className="font-medium text-blue-600 dark:text-blue-500 hover:underline"
-                      >
-                        <img src={deleted} alt="" />
-                      </button>
-                    </div>
-                  </td>
-                </tr>
-              ))}
+                    <button
+                      onClick={() => handleDeleteUser(user.id)}
+                      className="font-medium text-blue-600 dark:text-blue-500 hover:underline"
+                    >
+                      <img src={deleted} alt="" />
+                    </button>
+                  </div>
+                </td>
+              </tr>
+            ))}
           </tbody>
         </table>
       </div>
